refact: merge util.creation to core.creation and squash
